fix: move db upgrade logic to migration from service
