+++ tests/integration_test.rs
@@ -42,6 +42,7 @@ mod tests {
                 router::*,
             },
         },
+        migration,
         service::{self, alerts::scheduler::handlers::handle_triggers, search::SEARCH_SERVER},
     };
     use prost::Message;
@@ -128,7 +129,7 @@ mod tests {
         // init config
         config::init().await.unwrap();
         // init infra
-        service::init_db().await.unwrap();
+        migration::init_db().await.unwrap();
         infra::init().await.unwrap();
         // db migration steps, since it's separated out
         infra::table::migrate().await.unwrap();
